update calculate module reduce a operations stack to result
